refactor: enhance ActionsRetriever with comprehensive action analysis features
- Expanded the ActionsRetriever class to include new methods for retrieving and analyzing action data, such as actions, action_stats, temporal_patterns, and resource_impacts, providing a more flexible and detailed analytical framework.
- Introduced new data types like AgentActionData and enhanced ActionMetrics to encapsulate comprehensive metrics, including interaction rates and performance comparisons.
- Updated documentation throughout the ActionsRetriever class, improving clarity and usability with detailed descriptions, examples, and notes on analysis scopes.
- Streamlined the overall structure of action-related data types, ensuring consistency and clarity in how agent actions and interactions are represented and analyzed.

These changes significantly improve the framework for analyzing agent behaviors and interactions within the simulation environment, facilitating deeper insights and more effective decision-making analysis.
